Application admin vs student

+++ src/app/Controllers/ApplicationController.php
@@ -28,14 +28,34 @@ class ApplicationController extends BaseController
 
     public function index()
     {
-        // SELECT * FROM program;
-        $query = $this->db->query('SELECT program.program_num, program.name, application.status, application.app_num
-        FROM program JOIN application ON program.program_num = application.program_num;');
+        $user = sessionUser();
+        if(!$user) return $this->response->redirect(site_url('/login'));
+
+        $query;
+        $student_name;
+
+        if($user->User_Type == 'student'){
+            // SELECT * FROM program;
+            $sql = <<<SQL
+                SELECT program.program_num, program.name, application.status, application.app_num 
+                FROM program JOIN application ON application.program_num = program.program_num
+                WHERE application.UIN = $user->UIN; 
+            SQL;
+            $query = $this->db->query($sql);
+        }else{
+            $sql = <<<SQL
+                SELECT program.program_num, program.name, application.status, application.app_num, user.First_Name, user.Last_Name
+                FROM program JOIN application ON application.program_num = program.program_num 
+                JOIN user ON application.UIN = user.UIN
+            SQL;
+            $query = $this->db->query($sql);
+        }
 
         $data = [
             'page_title' => 'View Applications | TAMU CyberSec Center',
             // use getResultArray() on queries that return multiple rows
-            'applications' => $query->getResultArray()
+            'applications' => $query->getResultArray(),
+            'user' => $user
         ];
 
         return view('application/index', $data);
@@ -43,19 +63,35 @@ class ApplicationController extends BaseController
     
     public function create() {
         // select only programs that the user hasn't applied for already
-        // TODO: use session for UIN = 
-        $query = $this->db->query('SELECT * FROM program WHERE program_num NOT IN (SELECT program_num FROM application WHERE UIN = 230006744);');
+        $user = sessionUser();
+        if(!$user) return $this->response->redirect(site_url('/login'));
 
-        $data = [
-            'page_title' => 'Applications | TAMU CyberSec Center',
-            'programs' => $query->getResultArray()
-        ];
+        $query;
 
-        return view('application/create', $data);
+        if($user->hasPermission('student')){
+            $sql = <<<SQL
+                SELECT * FROM program WHERE program_num NOT IN
+                (SELECT program_num FROM application WHERE UIN = $user->UIN);
+            SQL;
+            $query = $this->db->query($sql);
+
+            $data = [
+                'page_title' => 'Applications | TAMU CyberSec Center',
+                'programs' => $query->getResultArray()
+            ];
+
+            return view('application/create', $data);
+        }
+        return $this->response->redirect(site_url('/application'));
+
     }
 
     public function new()
     {
+        $user = sessionUser();
+        if(!$user) return $this->response->redirect(site_url('/login'));
+        if($user->hasPermission('admin')) return $this->response->redirect(site_url('/application'));
+
         $method = $this->request->getMethod();
 
         if ($method == "post") {
@@ -66,15 +102,21 @@ class ApplicationController extends BaseController
                 return $this->response->redirect(site_url('program/'));
             }
 
-            // TODO: get UIN from session
-            $this->db->query('INSERT INTO application (program_num, UIN, uncom_cert, com_cert, purpose_statement)
-             VALUES(\''.$formData['program'].'\',230006744,\''.$formData['uncom_cert'].'\',\''.$formData['com_cert'].'\',\''.$formData['purpose_statement'].'\');');
+            $sql =  <<<SQL
+                INSERT INTO application (program_num, UIN, uncom_cert, com_cert, purpose_statement)
+                VALUES('{$formData['program']}',$user->UIN,'{$formData['uncom_cert']}','{$formData['com_cert']}','{$formData['purpose_statement']}');
+            SQL;
+            $this->db->query($sql);
         }
 
-        return $this->response->redirect(site_url('program/'));
+        return $this->response->redirect(site_url('/application'));
 	}
 
     public function edit($app_num, $program_num){
+        $user = sessionUser();
+        if(!$user) return $this->response->redirect(site_url('/login'));
+        if($user->hasPermission('admin')) return $this->response->redirect(site_url('/application'));
+
         $query = $this->db->query('SELECT * FROM application WHERE app_num = \''.$app_num.'\';');
         $programQuery = $this->db->query('SELECT name FROM program WHERE program_num = '.$program_num.';');
         $data = [
@@ -87,6 +129,10 @@ class ApplicationController extends BaseController
         return view('application/edit', $data);
     }
     public function update($app_num){
+        $user = sessionUser();
+        if(!$user) return $this->response->redirect(site_url('/login'));
+        if($user->hasPermission('admin')) return $this->response->redirect(site_url('/application'));
+
         $method = $this->request->getMethod();
         if($method == "post"){
             $formData = $this->request->getPost();
@@ -98,6 +144,10 @@ class ApplicationController extends BaseController
     }
 
 	public function delete($app_num){
+        $user = sessionUser();
+        if(!$user) return $this->response->redirect(site_url('/login'));
+        if($user->hasPermission('admin')) return $this->response->redirect(site_url('/application'));
+
 		$query = $this->db->query('SELECT * FROM application WHERE app_num = '.$app_num.';');
         $application = $query->getResultArray();
         $programQuery = $this->db->query('SELECT name FROM program WHERE program_num = '.$application[0]['program_num'].';');
@@ -112,6 +162,10 @@ class ApplicationController extends BaseController
 	}
 
 	public function destroy($id){
+        $user = sessionUser();
+        if(!$user) return $this->response->redirect(site_url('/login'));
+        if($user->hasPermission('admin')) return $this->response->redirect(site_url('/application'));
+
         if($id != null){
            $this->db->query('DELETE FROM application WHERE app_num = '.$id.';');
         }
